Broken down the API Specification into separate files for easier code understanding
